Move buildRegEx function to test/err.spec.ts
